iterate and test run on user shader

+++ Src/Projects/manager_PostProcessing/postprocesscontextdata.cxx
@@ -1036,7 +1036,7 @@ void PostProcessContextData::DrawHUDText(FBHUDTextElement *pRect, CFont *pFont, 
 
     }
 
-    const int numchars = strlen(buffer);
+    const int numchars = static_cast<int>(strlen(buffer));
 
     // draw a background
